requirements.txt
beautifulsoup4==4.6.0
Django==2.1.4
psycopg2==2.7.3.1
python-dateutil==2.6.1
requests==2.18.4